updated backlog page
