+++ tailwind.config.js
@@ -46,6 +46,10 @@ module.exports = {
         500: '#ecc94b',
         600: '#F8B500', // primary yellow
       },
+      green: {
+        100: '#D1FAE5',
+        600: '#059669',
+      }
     },
     spacing: {
       px: '1px',
